requirements.txt
Flask
gunicorn
webdriver-manager
requests
beautifulsoup4
pyyaml
validators
selenium
undetected-chromedriver
